Restyle 'add' link and switch icon font to svg icon system
Changed the 'add your quickstart or cartridge' link to be alongside
the rest of the site navigation. Also restyled it to be more button-
like.

As part of the restyling, I needed to add a new icon. The icon font
we had previously made that difficult, so I updated the icon system
to use an SVG based approach instead. The SVG based icon approach
is beginning to have more favor as a best practice over the icon
font approach for better maintainability and compatibility.

As part of the switch, I added a grunt task using the grunt-svgstore
plugin to concatenate a folder of individual icons into a single
file with each icon as a symbol. For future icon updates, it will
only require adding a new icon SVG to the folder, and rerunning the
grunt task.

I also added a template context variable to insert the svg into the
layout template, and wrote an icon macro into the layout template.

See http://css-tricks.com/svg-sprites-use-better-icon-fonts/ for a
detailed run down of the SVG icon system.